Add files via upload
This project is to develop a wheelchair control which is useful to the physically disabled person with his hand movement or his hand gesture recognition using MEMS technology. The use of powered wheelchair with high navigational intelligence is one of the great steps towards the integration of severely physically disabled people. Driving wheelchair in domestic environments is a difficult task for people with arm or hands impairments. The wheelchair is developed to overcome the above problem described above allowing the end users to just perform safe movements and accomplish some daily life important tasks.